added single source of FormFactorLookup helper structs
