+++ source/core/form_factor/lookup/FormFactorProduct.cpp
@@ -3,6 +3,7 @@
 
 #include <form_factor/lookup/FormFactorProduct.h>
 #include <form_factor/lookup/detail/FormFactorProductBase.h>
+#include <form_factor/lookup/detail/LookupHelpers.h>
 #include <form_factor/FormFactor.h>
 #include <constants/Constants.h>
 
@@ -10,13 +11,7 @@ using namespace ausaxs;
 using namespace ausaxs::form_factor;
 
 namespace {
-    struct RawFormFactorLookup {
-        static constexpr const FormFactor& get(form_factor_t type) {
-            return lookup::atomic::raw::get(type);
-        }
-    };
-
-    auto ff_table = lookup::detail::generate_atomic_table<RawFormFactorLookup>();
+    auto ff_table = lookup::detail::generate_atomic_table<lookup::detail::RawFormFactorLookup>();
 }
 
 const FormFactorProduct& form_factor::lookup::atomic::raw::get_product(unsigned int i, unsigned int j) noexcept {
