+++ source/core/form_factor/lookup/NormalizedExvFormFactorProduct.cpp
@@ -4,6 +4,7 @@
 #include <form_factor/lookup/NormalizedExvFormFactorProduct.h>
 #include <form_factor/lookup/CustomExvTable.h>
 #include <form_factor/lookup/detail/FormFactorProductBase.h>
+#include <form_factor/lookup/detail/LookupHelpers.h>
 #include <form_factor/NormalizedFormFactor.h>
 #include <settings/MoleculeSettings.h>
 #include <constants/Constants.h>
@@ -12,14 +13,8 @@ using namespace ausaxs;
 using namespace ausaxs::form_factor;
 
 namespace {
-    struct NormalizedFormFactorLookup {
-        static constexpr const NormalizedFormFactor& get(form_factor_t type) {
-            return lookup::atomic::normalized::get(type);
-        }
-    };
-
     auto ff_xx_default_table = lookup::detail::generate_exv_table<lookup::exv::table_t>(lookup::exv::standard);
-    auto ff_ax_default_table = lookup::detail::generate_cross_table<lookup::cross::table_t, NormalizedFormFactorLookup>(lookup::exv::standard);
+    auto ff_ax_default_table = lookup::detail::generate_cross_table<lookup::cross::table_t, lookup::detail::NormalizedFormFactorLookup>(lookup::exv::standard);
 }
 
 const NormalizedFormFactorProduct& form_factor::lookup::exv::normalized::get_product(unsigned int i, unsigned int j) {
@@ -40,6 +35,6 @@ const NormalizedFormFactorProduct& form_factor::lookup::cross::normalized::get_p
 const form_factor::lookup::cross::table_t& form_factor::lookup::cross::normalized::get_table() {
     return lookup::detail::get_exv_table_for_settings(
         ff_ax_default_table, lookup::exv::detail::custom_table_ax_normalized,
-        [](const form_factor::detail::ExvFormFactorSet& set) { return lookup::detail::generate_cross_table<lookup::cross::table_t, NormalizedFormFactorLookup>(set); }
+        [](const form_factor::detail::ExvFormFactorSet& set) { return lookup::detail::generate_cross_table<lookup::cross::table_t, lookup::detail::NormalizedFormFactorLookup>(set); }
     );
 }
